Change views inside client, add displayname attribute to models inside client, add feature in javascript
I modified views inside client in order to give them better look. Nextly I added feature in javascript which checking if forumpost belong to user or user is a admin and then decide if display delete and update icon in the views. Finally inside webApi I include UserDto when repo returning the forumposts which belong to category.